Move test/utils.py to helpers/test/utils.py
